+++ apps/backoffice/src/lib/be/services/business.ts
@@ -17,7 +17,11 @@ import { pipe } from "fp-ts/lib/function";
 import { NextRequest, NextResponse } from "next/server";
 import { BackOfficeUser, Institution } from "../../../../types/next-auth";
 import { getServiceList } from "./apim";
-import { IoServicesCmsClient, callIoServicesCms } from "./cms";
+import {
+  IoServicesCmsClient,
+  callIoServicesCms,
+  getServiceTopics
+} from "./cms";
 import {
   retrieveLifecycleServices,
   retrievePublicationServices
@@ -31,8 +35,10 @@ import {
 import {
   buildMissingService,
   reducePublicationServicesList,
-  toServiceListItem
+  toServiceListItem,
+  reduceServiceTopicsList
 } from "./utils";
+import { ServiceTopicList } from "@/generated/api/ServiceTopicList";
 
 type PathParameters = {
   serviceId?: string;
@@ -59,8 +65,14 @@ export const retrieveServiceList = async (
   pipe(
     getServiceList(userId, limit, offset, serviceId),
     TE.bindTo("apimServices"),
+    TE.bind("serviceTopicsMap", ({ apimServices }) =>
+      pipe(
+        TE.tryCatch(getServiceTopics, E.toError),
+        TE.map(({ topics }) => reduceServiceTopicsList(topics))
+      )
+    ),
     // get services from services-lifecycle cosmos containee and map to ServiceListItem
-    TE.bind("lifecycleServices", ({ apimServices }) =>
+    TE.bind("lifecycleServices", ({ apimServices, serviceTopicsMap }) =>
       pipe(
         apimServices.value
           ? apimServices.value.map(
@@ -68,7 +80,7 @@ export const retrieveServiceList = async (
             )
           : [],
         retrieveLifecycleServices,
-        TE.map(RA.map(toServiceListItem))
+        TE.map(RA.map(toServiceListItem(serviceTopicsMap)))
       )
     ),
     // get services from services-publication cosmos container
@@ -205,6 +217,10 @@ export async function forwardIoServicesCmsRequest<
     );
   }
 }
+
+export const retrieveServiceTopics = async (): Promise<ServiceTopicList> => {
+  return await getServiceTopics();
+};
 
 /**
  * SUBSCRIPTIONS